+++ validator/rpc/apimiddleware/endpoint_factory.go
@@ -19,7 +19,6 @@ func (*ValidatorEndpointFactory) Paths() []string {
 		"/eth/v1/keystores",
 		"/eth/v1/remotekeys",
 		"/eth/v1/validator/{pubkey}/feerecipient",
-		"/eth/v1/validator/{pubkey}/gas_limit",
 	}
 }
 
@@ -43,10 +42,6 @@ func (*ValidatorEndpointFactory) Create(path string) (*apimiddleware.Endpoint, e
 		endpoint.GetResponse = &GetFeeRecipientByPubkeyResponseJson{}
 		endpoint.PostRequest = &SetFeeRecipientByPubkeyRequestJson{}
 		endpoint.DeleteRequest = &DeleteFeeRecipientByPubkeyRequestJson{}
-	case "/eth/v1/validator/{pubkey}/gas_limit":
-		endpoint.GetResponse = &GetGasLimitResponseJson{}
-		endpoint.PostRequest = &SetGasLimitRequestJson{}
-		endpoint.DeleteRequest = &DeleteGasLimitRequestJson{}
 	default:
 		return nil, errors.New("invalid path")
 	}
