+++ validator/rpc/standard_api_test.go
@@ -15,9 +15,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	fieldparams "github.com/prysmaticlabs/prysm/v4/config/fieldparams"
-	"github.com/prysmaticlabs/prysm/v4/config/params"
 	validatorserviceconfig "github.com/prysmaticlabs/prysm/v4/config/validator/service"
-	"github.com/prysmaticlabs/prysm/v4/consensus-types/validator"
 	"github.com/prysmaticlabs/prysm/v4/crypto/bls"
 	"github.com/prysmaticlabs/prysm/v4/encoding/bytesutil"
 	ethpbservice "github.com/prysmaticlabs/prysm/v4/proto/eth/service"
@@ -40,7 +38,6 @@ import (
 	mocks "github.com/prysmaticlabs/prysm/v4/validator/testing"
 	keystorev4 "github.com/wealdtech/go-eth2-wallet-encryptor-keystorev4"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
 )
 
 func TestServer_ListKeystores(t *testing.T) {
@@ -1144,405 +1141,4 @@ func TestServer_DeleteFeeRecipientByPubkey_InvalidPubKey(t *testing.T) {
 
 	_, err := s.DeleteFeeRecipientByPubkey(ctx, req)
 	require.ErrorContains(t, "not a valid bls public key", err)
-}
-
-func TestServer_GetGasLimit(t *testing.T) {
-	ctx := context.Background()
-	byteval, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	byteval2, err2 := hexutil.Decode("0x1234567878903438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	require.NoError(t, err)
-	require.NoError(t, err2)
-
-	tests := []struct {
-		name   string
-		args   *validatorserviceconfig.ProposerSettings
-		pubkey [48]byte
-		want   uint64
-	}{
-		{
-			name: "ProposerSetting for specific pubkey exists",
-			args: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(byteval): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: 123456789},
-					},
-				},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: 987654321},
-				},
-			},
-			pubkey: bytesutil.ToBytes48(byteval),
-			want:   123456789,
-		},
-		{
-			name: "ProposerSetting for specific pubkey does not exist",
-			args: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(byteval): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: 123456789},
-					},
-				},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: 987654321},
-				},
-			},
-			// no settings for the following validator, so the gaslimit returned is the default value.
-			pubkey: bytesutil.ToBytes48(byteval2),
-			want:   987654321,
-		},
-		{
-			name:   "No proposerSetting at all",
-			args:   nil,
-			pubkey: bytesutil.ToBytes48(byteval),
-			want:   params.BeaconConfig().DefaultBuilderGasLimit,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			m := &mock.MockValidator{}
-			err := m.SetProposerSettings(ctx, tt.args)
-			require.NoError(t, err)
-			vs, err := client.NewValidatorService(ctx, &client.Config{
-				Validator: m,
-			})
-			require.NoError(t, err)
-			s := &Server{
-				validatorService: vs,
-			}
-			got, err := s.GetGasLimit(ctx, &ethpbservice.PubkeyRequest{Pubkey: tt.pubkey[:]})
-			require.NoError(t, err)
-			assert.Equal(t, tt.want, got.Data.GasLimit)
-		})
-	}
-}
-
-func TestServer_SetGasLimit(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	beaconClient := validatormock.NewMockValidatorClient(ctrl)
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-
-	pubkey1, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	pubkey2, err2 := hexutil.Decode("0xbedefeaa94e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2cdddddddddddddddddddddddd")
-	require.NoError(t, err)
-	require.NoError(t, err2)
-
-	type beaconResp struct {
-		resp  *eth.FeeRecipientByPubKeyResponse
-		error error
-	}
-
-	type want struct {
-		pubkey   []byte
-		gaslimit uint64
-	}
-
-	tests := []struct {
-		name             string
-		pubkey           []byte
-		newGasLimit      uint64
-		proposerSettings *validatorserviceconfig.ProposerSettings
-		w                []*want
-		beaconReturn     *beaconResp
-		wantErr          string
-	}{
-		{
-			name:             "ProposerSettings is nil",
-			pubkey:           pubkey1,
-			newGasLimit:      9999,
-			proposerSettings: nil,
-			wantErr:          "no proposer settings were found to update",
-		},
-		{
-			name:        "ProposerSettings.ProposeConfig is nil AND ProposerSettings.DefaultConfig is nil",
-			pubkey:      pubkey1,
-			newGasLimit: 9999,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: nil,
-				DefaultConfig: nil,
-			},
-			wantErr: "gas limit changes only apply when builder is enabled",
-		},
-		{
-			name:        "ProposerSettings.ProposeConfig is nil AND ProposerSettings.DefaultConfig.BuilderConfig is nil",
-			pubkey:      pubkey1,
-			newGasLimit: 9999,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: nil,
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					BuilderConfig: nil,
-				},
-			},
-			wantErr: "gas limit changes only apply when builder is enabled",
-		},
-		{
-			name:        "ProposerSettings.ProposeConfig is defined for pubkey, BuilderConfig is nil AND ProposerSettings.DefaultConfig is nil",
-			pubkey:      pubkey1,
-			newGasLimit: 9999,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(pubkey1): {
-						BuilderConfig: nil,
-					},
-				},
-				DefaultConfig: nil,
-			},
-			wantErr: "gas limit changes only apply when builder is enabled",
-		},
-		{
-			name:        "ProposerSettings.ProposeConfig is defined for pubkey, BuilderConfig is defined AND ProposerSettings.DefaultConfig is nil",
-			pubkey:      pubkey1,
-			newGasLimit: 9999,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(pubkey1): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{},
-					},
-				},
-				DefaultConfig: nil,
-			},
-			wantErr: "gas limit changes only apply when builder is enabled",
-		},
-		{
-			name:        "ProposerSettings.ProposeConfig is NOT defined for pubkey, BuilderConfig is defined AND ProposerSettings.DefaultConfig is nil",
-			pubkey:      pubkey2,
-			newGasLimit: 9999,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(pubkey2): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: 12345,
-						},
-					},
-				},
-				DefaultConfig: nil,
-			},
-			w: []*want{{
-				pubkey2,
-				9999,
-			},
-			},
-		},
-		{
-			name:        "ProposerSettings.ProposeConfig is defined for pubkey, BuilderConfig is nil AND ProposerSettings.DefaultConfig.BuilderConfig is defined",
-			pubkey:      pubkey1,
-			newGasLimit: 9999,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(pubkey2): {
-						BuilderConfig: nil,
-					},
-				},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					BuilderConfig: &validatorserviceconfig.BuilderConfig{
-						Enabled: true,
-					},
-				},
-			},
-			w: []*want{{
-				pubkey1,
-				9999,
-			},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			m := &mock.MockValidator{}
-			err := m.SetProposerSettings(ctx, tt.proposerSettings)
-			require.NoError(t, err)
-			validatorDB := dbtest.SetupDB(t, [][fieldparams.BLSPubkeyLength]byte{})
-			vs, err := client.NewValidatorService(ctx, &client.Config{
-				Validator: m,
-				ValDB:     validatorDB,
-			})
-			require.NoError(t, err)
-
-			s := &Server{
-				validatorService:          vs,
-				beaconNodeValidatorClient: beaconClient,
-				valDB:                     validatorDB,
-			}
-
-			if tt.beaconReturn != nil {
-				beaconClient.EXPECT().GetFeeRecipientByPubKey(
-					gomock.Any(),
-					gomock.Any(),
-				).Return(tt.beaconReturn.resp, tt.beaconReturn.error)
-			}
-
-			_, err = s.SetGasLimit(ctx, &ethpbservice.SetGasLimitRequest{Pubkey: tt.pubkey, GasLimit: tt.newGasLimit})
-			if tt.wantErr != "" {
-				require.ErrorContains(t, tt.wantErr, err)
-			} else {
-				require.NoError(t, err)
-				for _, w := range tt.w {
-					assert.Equal(t, w.gaslimit, uint64(s.validatorService.ProposerSettings().ProposeConfig[bytesutil.ToBytes48(w.pubkey)].BuilderConfig.GasLimit))
-				}
-			}
-		})
-	}
-}
-
-func TestServer_SetGasLimit_ValidatorServiceNil(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-
-	s := &Server{}
-
-	_, err := s.SetGasLimit(ctx, nil)
-	require.ErrorContains(t, "Validator service not ready", err)
-}
-
-func TestServer_SetGasLimit_InvalidPubKey(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-	s := &Server{
-		validatorService: &client.ValidatorService{},
-	}
-
-	req := &ethpbservice.SetGasLimitRequest{
-		Pubkey: []byte{},
-	}
-
-	_, err := s.SetGasLimit(ctx, req)
-	require.ErrorContains(t, "not a valid bls public key", err)
-}
-
-func TestServer_DeleteGasLimit(t *testing.T) {
-	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
-	pubkey1, err := hexutil.Decode("0xaf2e7ba294e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2c06bd3713cb442072ae591493")
-	pubkey2, err2 := hexutil.Decode("0xbedefeaa94e03438ea819bd4033c6c1bf6b04320ee2075b77273c08d02f8a61bcc303c2cdddddddddddddddddddddddd")
-	require.NoError(t, err)
-	require.NoError(t, err2)
-
-	// This test changes global default values, we do not want this to side-affect other
-	// tests, so store the origin global default and then restore after tests are done.
-	originBeaconChainGasLimit := params.BeaconConfig().DefaultBuilderGasLimit
-	defer func() {
-		params.BeaconConfig().DefaultBuilderGasLimit = originBeaconChainGasLimit
-	}()
-
-	globalDefaultGasLimit := validator.Uint64(0xbbdd)
-
-	type want struct {
-		pubkey   []byte
-		gaslimit validator.Uint64
-	}
-
-	tests := []struct {
-		name             string
-		pubkey           []byte
-		proposerSettings *validatorserviceconfig.ProposerSettings
-		wantError        error
-		w                []want
-	}{
-		{
-			name:   "delete existing gas limit with default config",
-			pubkey: pubkey1,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(pubkey1): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: validator.Uint64(987654321)},
-					},
-					bytesutil.ToBytes48(pubkey2): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: validator.Uint64(123456789)},
-					},
-				},
-				DefaultConfig: &validatorserviceconfig.ProposerOption{
-					BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: validator.Uint64(5555)},
-				},
-			},
-			wantError: nil,
-			w: []want{
-				{
-					pubkey: pubkey1,
-					// After deletion, use DefaultConfig.BuilderConfig.GasLimit.
-					gaslimit: validator.Uint64(5555),
-				},
-				{
-					pubkey:   pubkey2,
-					gaslimit: validator.Uint64(123456789),
-				},
-			},
-		},
-		{
-			name:   "delete existing gas limit with no default config",
-			pubkey: pubkey1,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(pubkey1): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: validator.Uint64(987654321)},
-					},
-					bytesutil.ToBytes48(pubkey2): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: validator.Uint64(123456789)},
-					},
-				},
-			},
-			wantError: nil,
-			w: []want{
-				{
-					pubkey: pubkey1,
-					// After deletion, use global default, because DefaultConfig is not set at all.
-					gaslimit: globalDefaultGasLimit,
-				},
-				{
-					pubkey:   pubkey2,
-					gaslimit: validator.Uint64(123456789),
-				},
-			},
-		},
-		{
-			name:   "delete nonexist gas limit",
-			pubkey: pubkey2,
-			proposerSettings: &validatorserviceconfig.ProposerSettings{
-				ProposeConfig: map[[48]byte]*validatorserviceconfig.ProposerOption{
-					bytesutil.ToBytes48(pubkey1): {
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{GasLimit: validator.Uint64(987654321)},
-					},
-				},
-			},
-			wantError: fmt.Errorf("%s", codes.NotFound.String()),
-			w: []want{
-				// pubkey1's gaslimit is unaffected
-				{
-					pubkey:   pubkey1,
-					gaslimit: validator.Uint64(987654321),
-				},
-			},
-		},
-		{
-			name:      "delete nonexist gas limit 2",
-			pubkey:    pubkey2,
-			wantError: fmt.Errorf("%s", codes.NotFound.String()),
-			w:         []want{},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			m := &mock.MockValidator{}
-			err := m.SetProposerSettings(ctx, tt.proposerSettings)
-			require.NoError(t, err)
-			validatorDB := dbtest.SetupDB(t, [][fieldparams.BLSPubkeyLength]byte{})
-			vs, err := client.NewValidatorService(ctx, &client.Config{
-				Validator: m,
-				ValDB:     validatorDB,
-			})
-			require.NoError(t, err)
-			s := &Server{
-				validatorService: vs,
-				valDB:            validatorDB,
-			}
-			// Set up global default value for builder gas limit.
-			params.BeaconConfig().DefaultBuilderGasLimit = uint64(globalDefaultGasLimit)
-			_, err = s.DeleteGasLimit(ctx, &ethpbservice.DeleteGasLimitRequest{Pubkey: tt.pubkey})
-			if tt.wantError != nil {
-				assert.ErrorContains(t, fmt.Sprintf("code = %s", tt.wantError.Error()), err)
-			} else {
-				require.NoError(t, err)
-			}
-			for _, w := range tt.w {
-				assert.Equal(t, w.gaslimit, s.validatorService.ProposerSettings().ProposeConfig[bytesutil.ToBytes48(w.pubkey)].BuilderConfig.GasLimit)
-			}
-		})
-	}
 }
